main: ButtonCallbackTank: Correct the max value to 200
The CircuitCube::BuildVelocityCommand() adds 55 to the value sent
to the CircuitCube. As the max value is 255, passing 255 to it results
in an overflow, resulting in no Cube (Tank) movement. Limit the value
to 200 so that in the end 255, the max value, is sent. In main loop() see the
for() loop in TRAIN_CONTROL mode which limits the max value to 200, too.

This is for non TRAIN_CONTROL mode (i.e. TRAIN_CONTROL define disabled).

Signed-off-by: Dirk Behme <dirk.behme@gmail.com>