LEGORemote: Wait for remote bluetooth being enabled
Switching the CircuitCube on, i.e. in bluetooth mode, it permanently waits
for pairing. In contrast to this, pressing the green button on the Lego
PoweredUp remote control has a timeout and stops waiting for pairing if not
paired after some seconds.

As the existing code here scans for the PoweredUp remote control only once
at startup, this code and the remote control have to be started (powered & green
button pressed) more or less at the same time. This makes handling slightly
unconvenient and might result in a non-paired remote control.

Improve this by adding an infinite loop which waits for the remote control
being paired successfully.

Signed-off-by: Dirk Behme <dirk.behme@gmail.com>