main: ButtonCallbackTank: Use stop switch for 3rd channel
Use the left and right stop button to control the 3rd channel of the
CircuitCube. As we are using the RELEASED event to set the channels
back to 0, we don't need the stop buttons for the same. So we can use
the stop buttons for other actions.

In this change the left stop button is used to enable the 3rd CircuitCube
output, while the right stop button disables it. This can be used, e.g.
to control a light attached to this channel.

This is for non TRAIN_CONTROL mode (i.e. TRAIN_CONTROL define disabled).

Signed-off-by: Dirk Behme <dirk.behme@gmail.com>